add a button logout for the navbarusers
