+++ client/src/components/Navbar/NavBarAdmin.tsx
@@ -5,19 +5,22 @@ import IconButton from '@mui/material/IconButton';
 import Typography from '@mui/material/Typography';
 import Container from '@mui/material/Container';
 import Avatar from '@mui/material/Avatar';
-import { Link as MUILink } from '@mui/material';
 import { Link as RouterLink } from 'react-router-dom';
-import logo from '../assets/logopmp.png';
-
-const pages = [
-  { content: 'Gestion', to: '/itemmanagement' },
-  { content: 'Produits', to: '/product/view' },
-  { content: 'Marque', to: '/brand/view' }
-];
+import logo from '../../assets/logopmp.png';
+import { useAuth } from '../../context/AuthContext';
+import LogoutIcon from '@mui/icons-material/Logout';
 
 export default function NavBAr() {
+  const { logout } = useAuth();
   return (
-    <AppBar position="static" sx={{ backgroundColor: '#E3AB44' }}>
+    <AppBar
+      position="static"
+      sx={{
+        backgroundColor: '#E3AB44',
+        display: 'flex',
+        justifyContent: 'center'
+      }}
+    >
       <Container maxWidth="xl">
         <Toolbar disableGutters>
           <Box
@@ -43,29 +46,26 @@ export default function NavBAr() {
             </Typography>
           </Box>
 
-          <Box sx={{ flexGrow: 1, display: 'flex', justifyContent: 'right' }}>
-            {pages.map((page) => (
-              <MUILink
-                to={page.to}
-                key={page.content}
-                component={RouterLink}
-                sx={{
-                  my: 2,
-                  color: 'black',
-                  marginRight: 3,
-                  textTransform: 'uppercase',
-                  textDecoration: 'none',
-                  fontFamily: "'Roboto', sans-serif"
-                }}
-              >
-                {page.content}
-              </MUILink>
-            ))}
-          </Box>
-          <Box sx={{ flexGrow: 0 }}>
-            <IconButton sx={{ p: 0 }}>
+          <Box
+            sx={{
+              display: 'flex',
+              alignItems: 'center',
+              justifyContent: 'flex-end',
+              width: '100%',
+              gap: 1
+            }}
+          >
+            <IconButton>
               <Avatar alt="A" src="/static/images/avatar/2.jpg" />
             </IconButton>
+            <LogoutIcon
+              onClick={logout}
+              sx={{
+                cursor: 'pointer',
+                color: 'black',
+                marginLeft: '2'
+              }}
+            />
           </Box>
         </Toolbar>
       </Container>
